created inputvalidation to refactor the code
